router/requirements.txt
# 统一AI服务路由系统依赖
# 整合final_review_agent、thesis_agent、web_agent三个服务

# FastAPI核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据验证和序列化
pydantic==2.5.0

# HTTP客户端和异步支持
httpx==0.25.2
aiofiles==23.2.1

# OpenAI API客户端
openai==1.3.7

# 环境变量管理
python-dotenv==1.0.0

# 文件上传和表单处理
python-multipart==0.0.6

# HTTP请求库
requests==2.31.0

# 日期时间处理
python-dateutil==2.8.2

# 开发和测试工具
pytest==7.4.3
pytest-asyncio==0.21.1

# 性能监控（可选）
prometheus-client==0.19.0

# 生产环境WSGI服务器（可选）
# gunicorn==21.2.0
